+++ src/routes/editor/+page.svelte
@@ -2,44 +2,156 @@
     import { onMount, onDestroy } from "svelte";
     import { globalComponentCollectionStore, globalEditorPreferencesStore, globalVisibilityStore,
         globalRightPanelContentStore, globalLeftPanelContentStore  } from "../globals/globalstores.js";
-    import { PanelDisplayStyles, MenuLocations } from "../globals/globalconstants.js";
+    import { PanelDisplayStyles, MenuLocations, ScreenSizePx } from "../globals/globalconstants.js";
     import jQuery from "jquery";
 
     let globalComponentCollection = $globalComponentCollectionStore ?? [{}] ;
-    let globalEditorPreferences = $globalEditorPreferencesStore ?? {};
-
     $: globalComponentCollectionStore.set(globalComponentCollection);
+
+    let globalEditorPreferences = $globalEditorPreferencesStore ?? {};
     $: globalEditorPreferencesStore.set(globalEditorPreferences);
 
-    //Editor Theme settings
-    let editorTheme = globalEditorPreferences.editorTheme;
+    /// update bootstrap breakpoints each time editor width is changed
+    $: updateBootstrapBreakPoints($globalEditorPreferencesStore.editorData.editorWidth);
 
-    function setMenuLocation(location){
-        // console.log("menuLocation:" + location);
-        globalEditorPreferences.menuLocation = location;
+
+
+    let bsWidthSm = "100%";
+    let bsWidthMd = "100%";
+    let bsWidthLg = "100%";
+    let bsWidthXl = "100%";
+    let bsWidthXxl = "100%";
+    let bsWidthContainer = "100%";
+
+    /**
+     * Width of screen
+     */
+    let outerWidth = 0;
+
+    let editorPanelWidth = 0;
+    let editorPanelHeight = 0;
+
+    let testPanelWidth = 0;
+
+    /**
+     * Updates bootstrap breakpoints; which is specifically derived from original for Editor, located at "editorcss.css" imported below.
+     *
+     * This is crucial since we make our editor supports bootstrap natively
+     * @param {string} width Width of Editor
+     */
+    function updateBootstrapBreakPoints(width){
+
+        var editorWidth = Math.min(parseInt(width.replace("px","")), parseInt(editorPanelWidth.toString().replace("px","")));
+
+
+        var xxl = 1320;
+        var xl = 1140;
+        var lg = 960;
+        var md = 720;
+        var sm = 540;
+
+        if(width == ScreenSizePx.DESKTOP && $globalEditorPreferencesStore.editorData.fullWidth == true){
+            // bsWidthXxl = "100%";
+            // bsWidthXl = "100%";
+            // bsWidthLg = "100%";
+            // bsWidthMd = "100%";
+            // bsWidthSm = "100%";
+            // bsWidthContainer = "100%";
+            var value = Math.max(xxl, editorWidth);
+                bsWidthXxl = value+"px";
+                bsWidthXl = value+"px";
+                bsWidthLg = value+"px";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+        }else{
+
+            if(editorWidth > 1400){
+                // var newWidth = Math.round(editorWidth / xxl);
+                // var value = Math.min(xxl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+                var value = Math.max(xxl, editorWidth);
+                bsWidthXxl = value+"px";
+                bsWidthXl = value+"px";
+                bsWidthLg = value+"px";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+            }else if( editorWidth <= 1400 && editorWidth > 1200){
+                // var newWidth = Math.round(editorWidth / xl);
+                // var value = Math.min(xl, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+                var value = Math.max(xl, editorWidth);
+                bsWidthXxl = "100%";
+                bsWidthXl = value+"px";
+                bsWidthLg = value+"px";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+            }else if( editorWidth <= 1200 && editorWidth > 992){
+                // var newWidth = Math.round(editorWidth / lg);
+                // var value = Math.min(lg, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+                var value = Math.max(lg, editorWidth);
+                bsWidthXxl = "100%";
+                bsWidthXl = "100%";
+                bsWidthLg = value+"px";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+            }else if( editorWidth <= 992 && editorWidth > 768){
+                // var newWidth = Math.round(editorWidth / md);
+                // var value = Math.min(md, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+                var value = Math.max(md, editorWidth);
+                bsWidthXxl = "100%";
+                bsWidthXl = "100%";
+                bsWidthLg = "100%";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+            }else if( editorWidth <= 768 && editorWidth > 576){
+                // var newWidth = Math.round(editorWidth / sm);
+                // var value = Math.min(sm, newWidth % 2 == 0 ? newWidth : newWidth + 1 ); //Resolution must be even.
+                // console.log("editorWidth <= 768 && editorWidth > 576 ||=> newWidth: "+newWidth.toString()+" | sm: "+sm.toString());
+                console.log("editorWidth <= 768 && editorWidth > 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
+                var value = Math.max(sm, editorWidth);
+                bsWidthXxl = "100%";
+                bsWidthXl = "100%";
+                bsWidthLg = "100%";
+                bsWidthMd = "100%";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+            }else if( editorWidth <= 576){
+                console.log("editorWidth <= 576 ||=> editorWidth: "+editorWidth.toString()+" | sm: "+sm.toString());
+                var value = editorWidth;
+                bsWidthXxl = value+"px";
+                bsWidthXl = value+"px";
+                bsWidthLg = value+"px";
+                bsWidthMd = value+"px";
+                bsWidthSm = value+"px";
+                bsWidthContainer = value+"px";
+                // bsWidthXxl = "100%";
+                // bsWidthXl = "100%";
+                // bsWidthLg = "100%";
+                // bsWidthMd = "100%";
+                // bsWidthSm = "100%";
+                // bsWidthContainer = "100%";
+            }
+
+        }
+
+                //////////////////////////////////////////////////////////////////////////////////////
+                //////// TODO: CONTAINER OLCULERI TAMAM AMA YINEDE HER BIR COL ICIN ORAN HESAPLAMAM LAZIM. BOOTSTRAP BAŞTAN BAŞLANACAK
+                //////////////////////////////////////////////////////////////////////////////////////
+
     }
 
 
-    // let editorData = $globalEditorPreferencesStore.editorData;
-    // let editorWidth = $globalEditorPreferencesStore.editorData.editorWidth;
-    // let editorScreen = $globalEditorPreferencesStore.editorData.editorScreen;
-    // let fullWidth = $globalEditorPreferencesStore.editorData.fullWidth;
-    let editorWidth = globalEditorPreferences.editorData.editorWidth;
-    let editorScreen = globalEditorPreferences.editorData.editorScreen;
-    let fullWidth = globalEditorPreferences.editorData.fullWidth;
 
-    // jQuery("*")
-    //     .on("mouseenter", (e) => {
-    //         var target = e.target;
-    //         jQuery(target).addClass("hovered");
-
-    //     }).on("mouseleave", (e) => {
-    //         var target = e.target;
-    //         jQuery(target).removeClass("hovered");
-    //     });
-
+    /**
+     * Sets selected element in editor.
+     * @param {string | undefined} uuid Unique identity string used in "globalComponentCollectionStore"
+     */
     function setSelectedElement(uuid){
         globalEditorPreferences.info = uuid;
+
     }
 
     onMount(() => {
@@ -48,7 +160,6 @@
         .on("mouseover", (e) => {
             var target = e.target;
             jQuery(target).addClass("hovered");
-            // console.log("mouseenter event");
             setSelectedElement(jQuery(target).attr("class"));
 
         }).on("mouseout", (e) => {
@@ -61,13 +172,21 @@
 
 </script>
 
-<input type="hidden" class="hovered" />
+<svelte:window bind:outerWidth={outerWidth}/>
 
-<div id="editorInnerPanel" class="clearfix" style='
-    --editorWidth:{$globalEditorPreferencesStore.editorData.editorWidth}; 
+
+<div id="editorInnerPanel" bind:clientWidth={editorPanelWidth} bind:clientHeight={editorPanelHeight} class="clearfix" style='
+    --editorWidth:{$globalEditorPreferencesStore.editorData.editorWidth};
+    --bsWidthSm:{bsWidthSm};
+    --bsWidthMd:{bsWidthMd};
+    --bsWidthLg:{bsWidthLg};
+    --bsWidthXl:{bsWidthXl};
+    --bsWidthXxl:{bsWidthXxl};
+    --bsWidthContainer:{bsWidthContainer};
 
 '
 >
+    <input type="hidden" class="hovered col" />
 
 
     <div class="container-fluid d-inline">
@@ -75,12 +194,23 @@
         <button class="btn btn-info">Test Button</button>
     </div>
 
-    <div class="container-fluid row">
-        <div class="col-xl-6 col-lg-6 col-md-12 col-sm-12 col-xs-12">
+    <div class="my-3">
+        --editorWidth:{$globalEditorPreferencesStore.editorData.editorWidth};
+        --bsWidthSm:{bsWidthSm};
+        --bsWidthMd:{bsWidthMd};
+        --bsWidthLg:{bsWidthLg};
+        --bsWidthXl:{bsWidthXl};
+        --bsWidthXxl:{bsWidthXxl};
+        --bsWidthContainer:{bsWidthContainer};
+        --testPanelWidth:{testPanelWidth};
+    </div>
+
+    <div class="container row">
+        <div class="col-xl-6 col-lg-6 col-md-12 col-sm-12" bind:clientWidth={testPanelWidth} id="cntrtest">
             <h4>Global Editor Preferences</h4>
             <pre>{JSON.stringify($globalEditorPreferencesStore, null, 2)}</pre>
         </div>
-        <div class="col-xl-6 col-lg-6 col-md-12 col-sm-12 col-xs-12">
+        <div class="col-xl-6 col-lg-6 col-md-12 col-sm-12">
             <h4>Global Visibility</h4>
             <pre>{JSON.stringify($globalVisibilityStore, null, 2)}</pre>
             <h4>Global Right Panel Content Store</h4>
@@ -194,17 +324,20 @@
     &nbsp;*&nbsp;<br/>
     &nbsp;*&nbsp;<br/>
     &nbsp;*&nbsp;<br/>
-    
+
 
 
 </div>
 
-    
+
 </div> <!-- editorInnerPanel -->
 
 
 
 <style>
+
+    /* @import "editorcss.css"; */
+
     #editorInnerPanel{
         width: var(--editorWidth) !important;
         vertical-align: top;
@@ -219,10 +352,113 @@
         flex-wrap: wrap; */
     }
 
-    .hovered{
+    /* .hovered{
         border: 2px solid aqua;
-        
+    } */
+
+    .container{
+    max-width: var(--bsWidthContainer) !important;
     }
+    .container-sm{
+        max-width: var(--bsWidthSm) !important;
+    }
+    .container-md{
+        max-width: var(--bsWidthMd) !important;
+    }
+    .container-lg{
+        max-width: var(--bsWidthLg) !important;
+    }
+    .container-xl{
+        max-width: var(--bsWidthXl) !important;
+    }
+    .container-xxl{
+        max-width: var(--bsWidthXxl) !important;
+    }
+
+  .col-sm {
+    flex: 1 0 0%;
+  }
+  .row-cols-sm-auto > * {
+    flex: 0 0 auto;
+    width: auto;
+  }
+  .row-cols-sm-1 > * {
+    flex: 0 0 auto;
+    width: 100%;
+  }
+  .row-cols-sm-2 > * {
+    flex: 0 0 auto;
+    width: 50%;
+  }
+  .row-cols-sm-3 > * {
+    flex: 0 0 auto;
+    width: 33.3333333333%;
+  }
+  .row-cols-sm-4 > * {
+    flex: 0 0 auto;
+    width: 25%;
+  }
+  .row-cols-sm-5 > * {
+    flex: 0 0 auto;
+    width: 20%;
+  }
+  .row-cols-sm-6 > * {
+    flex: 0 0 auto;
+    width: 16.6666666667%;
+  }
+  .col-sm-auto {
+    flex: 0 0 auto;
+    width: auto;
+  }
+  .col-sm-1 {
+    flex: 0 0 auto;
+    width: 8.33333333%;
+  }
+  .col-sm-2 {
+    flex: 0 0 auto;
+    width: 16.66666667%;
+  }
+  .col-sm-3 {
+    flex: 0 0 auto;
+    width: 25%;
+  }
+  .col-sm-4 {
+    flex: 0 0 auto;
+    width: 33.33333333%;
+  }
+  .col-sm-5 {
+    flex: 0 0 auto;
+    width: 41.66666667%;
+  }
+  .col-sm-6 {
+    flex: 0 0 auto;
+    width: 50%;
+  }
+  .col-sm-7 {
+    flex: 0 0 auto;
+    width: 58.33333333%;
+  }
+  .col-sm-8 {
+    flex: 0 0 auto;
+    width: 66.66666667%;
+  }
+  .col-sm-9 {
+    flex: 0 0 auto;
+    width: 75%;
+  }
+  .col-sm-10 {
+    flex: 0 0 auto;
+    width: 83.33333333%;
+  }
+  .col-sm-11 {
+    flex: 0 0 auto;
+    width: 91.66666667%;
+  }
+  .col-sm-12 {
+    flex: 0 0 auto;
+    width: 100%;
+  }
+
 
     /* sm: 540px,
   md: 720px,
@@ -231,23 +467,7 @@
   xxl: 1320px */
 
 
-    @media (min-width: 400px) and (max-width: 599px) {
-        .editorInnerPanel .col {
-            width: 50%;
-        }
-    }
 
-    @media (min-width: 600px) and (max-width: 799px) {
-        .editorInnerPanel .col {
-            width: 33.33%;
-        }
-    }
-
-    @media (min-width: 800px) {
-        .editorInnerPanel .col {
-            width: 25%;
-        }
-    }
 
 
 </style>